requirements.txt
llama-index
streamlit==1.31.1
python-dotenv==1.0.0
openai==1.12.0
pypdf==3.17.1
docx2txt==0.8
pandas==2.1.3
numpy==1.24.3
chromadb==0.4.18
transformers==4.36.2
huggingface-hub==0.20.3
requests==2.31.0
torch==2.1.2
pyyaml==6.0.1
spacy==3.7.4
pyspellchecker==0.7.2
scikit-learn
Pillow
PyMuPDF
pdf2image
pytesseract
easyocr
tabula-py
llama-index-retrievers-bm25
tiktoken
sentence-transformers
plotly